requirements.txt
asana==0.10.3
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
idna==2.10
oauthlib==3.1.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==1.9.6
urllib3==1.26.7
